Add core functionality for audio processing, authentication, project management, and version control
- Implemented audio processing logic in ProcAudio.cpp.
- Added authentication handling in ProcAuth.cpp, including sign-in and session restoration.
- Developed project job management in ProcProjJobs.cpp, enabling project activation and branch history retrieval.
- Created project handling functions in ProcProject.cpp, allowing project selection and version restoration.
- Introduced version job processing in ProcVerJobs.cpp, facilitating version pushing and restoring.
- Enhanced session caching and error handling throughout the application.
