+++ plugin/stemhub/CMakeLists.txt
@@ -45,8 +45,11 @@ target_sources(stemhub
         Source/src/ui/PluginEditor.cpp
         Source/src/ui/DashboardView.cpp
         Source/src/ui/LoginView.cpp
-        Source/src/application/PluginProcessor.Audio.cpp
-        Source/src/application/PluginProcessorStateMachine.cpp
+        Source/src/application/ProcAudio.cpp
+        Source/src/application/ProcAuth.cpp
+        Source/src/application/ProcProject.cpp
+        Source/src/application/ProcProjJobs.cpp
+        Source/src/application/ProcVerJobs.cpp
         Source/src/application/PluginProcessor.cpp
         Source/src/application/ProjectFileService.cpp
         Source/src/application/SessionCache.cpp
